Feat: Update Squad.java

+++ api/src/main/java/com/khali/api3/domain/squad/Squad.java
@@ -26,7 +26,7 @@ public class Squad {
     @Column(unique=true)
     private String acronym;
 
-    private int managerId;
+    private Long managerId;
 
     private String managerName;
